Move sofort notification handler to controller
