+++ tests/Unit/Routes/SofortNotificationRequestTest.php
@@ -6,7 +6,7 @@ namespace WMDE\Fundraising\Frontend\Tests\Unit\Routes;
 
 use DateTime;
 use PHPUnit\Framework\TestCase;
-use WMDE\Fundraising\Frontend\App\RouteHandlers\SofortNotificationHandler;
+use WMDE\Fundraising\Frontend\App\Controllers\SofortNotificationController;
 use WMDE\Fundraising\PaymentContext\RequestModel\SofortNotificationRequest;
 
 /**
@@ -19,7 +19,7 @@ class SofortNotificationRequestTest extends TestCase {
 			. '<status_notification><transaction>555-777</transaction>'
 			. '<time>2010-04-14T19:01:08+02:00</time>'
 			. '</status_notification>';
-		$request = SofortNotificationHandler::fromUseCaseRequestFromRequestContent( $content );
+		$request = SofortNotificationController::fromUseCaseRequestFromRequestContent( $content );
 
 		$this->assertInstanceOf( SofortNotificationRequest::class, $request );
 		$this->assertSame( '555-777', $request->getTransactionId() );
@@ -28,7 +28,7 @@ class SofortNotificationRequestTest extends TestCase {
 	}
 
 	public function testAbsurdRequest_nullIsReturned(): void {
-		$request = SofortNotificationHandler::fromUseCaseRequestFromRequestContent( 'fff' );
+		$request = SofortNotificationController::fromUseCaseRequestFromRequestContent( 'fff' );
 
 		$this->assertNull( $request );
 	}
@@ -38,7 +38,7 @@ class SofortNotificationRequestTest extends TestCase {
 			. '<status_notification>'
 			. '<time>2013-06-25T11:04:03+05:00</time>'
 			. '</status_notification>';
-		$request = SofortNotificationHandler::fromUseCaseRequestFromRequestContent( $content );
+		$request = SofortNotificationController::fromUseCaseRequestFromRequestContent( $content );
 
 		$this->assertNull( $request );
 	}
@@ -48,7 +48,7 @@ class SofortNotificationRequestTest extends TestCase {
 			. '<status_notification><transaction>555-777</transaction>'
 			. '<time>now</time>'
 			. '</status_notification>';
-		$request = SofortNotificationHandler::fromUseCaseRequestFromRequestContent( $content );
+		$request = SofortNotificationController::fromUseCaseRequestFromRequestContent( $content );
 
 		$this->assertNull( $request );
 	}
